src/requirements.txt
Django
djangorestframework
pymongo
Pygments
Markdown
pytz